Proper color of refname labels with cytobands on
